+++ src/lob.rs
@@ -87,6 +87,11 @@ impl PendingLobData {
             }))
         }
     }
+
+    /// Returns the locator.
+    pub(crate) fn take_locator(&mut self) -> Vec<u8> {
+        std::mem::take(&mut self.locator)
+    }
 }
 
 impl fmt::Debug for Lob {
@@ -194,6 +199,43 @@ impl Lob {
         Ok(())
     }
 
+    /// Creates an empty temporary LOB.
+    pub(crate) fn create_temp(
+        client_ref: ClientRef,
+        db_type: &'static DbType,
+    ) -> Result<Lob, Error> {
+        if db_type != &crate::DB_TYPE_BLOB
+            && db_type != &crate::DB_TYPE_CLOB
+            && db_type != &crate::DB_TYPE_NCLOB
+        {
+            return Err(Error::unsupported_db_type(db_type));
+        }
+        let locator = vec![0; 40];
+        let mut message = {
+            let mut client = client_ref.lock().unwrap();
+            let mut message = LobOpMessage::new(
+                &locator,
+                LobOp::CreateTemp {
+                    ora_type_num: db_type.ora_type_num,
+                    csfrm: db_type.csfrm,
+                },
+            );
+
+            client.process_message(&mut message)?;
+            message
+        };
+
+        let locator = message.take_returned_locator().unwrap();
+        Ok(Self {
+            client_ref,
+            locator,
+            db_type,
+            size: Some(0),
+            chunk_size: None,
+            offset: 1,
+        })
+    }
+
     /// Creates a LOB from its internal data.
     pub(crate) fn new(client_ref: ClientRef, data: PendingLobData) -> Self {
         Self {
@@ -258,6 +300,13 @@ impl Lob {
             self.offset = new_size + 1;
         }
         Ok(())
+    }
+}
+
+impl Drop for Lob {
+    fn drop(&mut self) {
+        let locator = std::mem::take(&mut self.locator);
+        self.client_ref.lock().unwrap().add_lob_to_close(locator);
     }
 }